[backend] Refact
