+++ openbas-api/src/main/java/io/openbas/rest/inject/service/InjectAssistantService.java
@@ -4,19 +4,11 @@ import static java.util.Collections.emptyList;
 
 import io.openbas.database.helper.InjectorContractRepositoryHelper;
 import io.openbas.database.model.*;
-import io.openbas.database.repository.InjectRepository;
-import io.openbas.database.repository.InjectorContractRepository;
 import io.openbas.injectors.manual.ManualContract;
-import io.openbas.rest.attack_pattern.service.AttackPatternService;
-import io.openbas.rest.exception.UnprocessableContentException;
-import io.openbas.rest.inject.form.InjectAssistantInput;
 import io.openbas.rest.injector_contract.InjectorContractContentUtils;
 import io.openbas.rest.injector_contract.InjectorContractService;
-import io.openbas.service.AssetGroupService;
-import io.openbas.service.EndpointService;
 import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
@@ -27,61 +19,10 @@ import org.springframework.validation.annotation.Validated;
 @Validated
 public class InjectAssistantService {
 
-  private final AssetGroupService assetGroupService;
-  private final EndpointService endpointService;
   private final InjectorContractService injectorContractService;
-  private final AttackPatternService attackPatternService;
-
-  private final InjectRepository injectRepository;
-  private final InjectorContractRepository injectorContractRepository;
   private final InjectorContractRepositoryHelper injectorContractRepositoryHelper;
 
   private InjectorContract manualInjectorContract = null;
-
-  /**
-   * Generate injects for a given scenario based on the provided input.
-   *
-   * @param scenario the scenario for which injects are generated
-   * @param input the input containing details for inject generation, such as attack pattern IDs,
-   *     asset IDs, asset group IDs, and the number of injects per TTP
-   * @return a list of generated injects
-   */
-  public Set<Inject> generateInjectsForScenario(Scenario scenario, InjectAssistantInput input) {
-    if (input.getInjectByTTPNumber() > 5) {
-      throw new UnsupportedOperationException(
-          "Number of inject by ttp must be less than or equal to 5");
-    }
-    List<Endpoint> endpoints = this.endpointService.endpoints(input.getAssetIds());
-    List<AssetGroup> assetGroups = this.assetGroupService.assetGroups(input.getAssetGroupIds());
-
-    Map<AssetGroup, List<Endpoint>> assetsFromGroupMap =
-        assetGroupService.assetsFromAssetGroupMap(assetGroups);
-
-    // Process injects generation for each attack pattern
-    Set<Inject> injects = new HashSet<>();
-    input
-        .getAttackPatternIds()
-        .forEach(
-            attackPatternId -> {
-              try {
-                Set<Inject> injectsToAdd =
-                    this.generateInjectsByTTP(
-                        attackPatternId,
-                        endpoints,
-                        assetsFromGroupMap,
-                        input.getInjectByTTPNumber());
-                injects.addAll(injectsToAdd);
-              } catch (UnprocessableContentException e) {
-                throw new UnsupportedOperationException(e);
-              }
-            });
-
-    for (Inject inject : injects) {
-      inject.setScenario(scenario);
-    }
-
-    return this.injectRepository.saveAll(injects).stream().collect(Collectors.toSet());
-  }
 
   /**
    * Builds an Inject object based on the provided InjectorContract, title, description and enabled
@@ -92,7 +33,7 @@ public class InjectAssistantService {
    * @param enabled indicates whether the Inject is enabled or not
    * @return the inject object built
    */
-  private Inject buildInject(
+  public Inject buildInject(
       InjectorContract injectorContract, String title, String description, Boolean enabled) {
     Inject inject = new Inject();
     inject.setTitle(title);
@@ -113,7 +54,7 @@ public class InjectAssistantService {
    * @param name the AttackPattern or Vulnerability associated with the Inject
    * @return the built Inject object
    */
-  private Inject buildTechnicalInjectFromInjectorContract(
+  public Inject buildTechnicalInjectFromInjectorContract(
       InjectorContract injectorContract, String identifier, String name) {
     return buildInject(
         injectorContract,
@@ -131,7 +72,7 @@ public class InjectAssistantService {
    * @param architecture the architecture to specify in the title and description of the Inject
    * @return the built manual Inject object
    */
-  private Inject buildManualInject(String identifier, String platform, String architecture) {
+  public Inject buildManualInject(String identifier, String platform, String architecture) {
     if (manualInjectorContract == null) {
       manualInjectorContract =
           this.injectorContractService.injectorContract(ManualContract.MANUAL_DEFAULT);
@@ -140,7 +81,7 @@ public class InjectAssistantService {
         manualInjectorContract,
         String.format("[%s] Placeholder - %s %s", identifier, platform, architecture),
         String.format(
-            "This placeholder is disabled because the TTP %s with platform %s and architecture %s is currently not covered. Please create the payloads for the missing TTP.",
+            "This placeholder is disabled because the AttackPattern %s with platform %s and architecture %s is currently not covered. Please create the payloads for the missing AttackPattern.",
             identifier, platform, architecture),
         false);
   }
@@ -178,17 +119,17 @@ public class InjectAssistantService {
       Map<String, List<Endpoint>> manualEndpoints) {}
 
   /**
-   * Get the injector contract for assets and TTP.
+   * Get the injector contract for assets and AttackPattern.
    *
    * @param attackPattern the attack pattern for which the injector contract need to match
-   * @param injectsPerTTP the maximum number of injector contracts to return
+   * @param injectsPerAttackPattern the maximum number of injector contracts to return
    * @param endpoints the list of endpoints to consider for the injector contract
    * @return a ContractResultForEndpoints containing the matched injector contracts with their
    *     endpoints, and the map of platform architecture pairs with the endpoints for those that
    *     didn't find injectorContract
    */
-  private ContractResultForEndpoints getInjectorContractForAssetsAndTTP(
-      AttackPattern attackPattern, Integer injectsPerTTP, List<Endpoint> endpoints) {
+  private ContractResultForEndpoints getInjectorContractForAssetsAndAttackPattern(
+      AttackPattern attackPattern, Integer injectsPerAttackPattern, List<Endpoint> endpoints) {
     Map<InjectorContract, List<Endpoint>> contractEndpointsMap = new HashMap<>();
     Map<String, List<Endpoint>> manualEndpoints = new HashMap<>();
 
@@ -199,7 +140,9 @@ public class InjectAssistantService {
     // Try to find injectors contract covering all platform-architecture pairs at once
     List<InjectorContract> injectorContracts =
         this.injectorContractRepositoryHelper.searchInjectorContractsByAttackPatternAndEnvironment(
-            attackPattern.getExternalId(), groupedAssets.keySet().stream().toList(), injectsPerTTP);
+            attackPattern.getExternalId(),
+            groupedAssets.keySet().stream().toList(),
+            injectsPerAttackPattern);
 
     if (!injectorContracts.isEmpty()) {
       injectorContracts.forEach(ic -> contractEndpointsMap.put(ic, endpoints));
@@ -213,7 +156,7 @@ public class InjectAssistantService {
                     .searchInjectorContractsByAttackPatternAndEnvironment(
                         attackPattern.getExternalId(),
                         List.of(platformArchitecture),
-                        injectsPerTTP);
+                        injectsPerAttackPattern);
 
             // Else take the manual injectorContract
             if (injectorContractsForGroup.isEmpty()) {
@@ -232,13 +175,13 @@ public class InjectAssistantService {
    * @param knownInjectorContracts the list of known injector contracts to search from
    * @param platformArchitecturePairs the list of platform-architecture pairs to filter the
    *     contracts
-   * @param injectsPerTTP the maximum number of injector contracts to return
+   * @param injectsPerAttackPattern the maximum number of injector contracts to return
    * @return a list of InjectorContract objects that match the search criteria
    */
   private List<InjectorContract> findInjectorContracts(
       List<InjectorContract> knownInjectorContracts,
       List<String> platformArchitecturePairs,
-      Integer injectsPerTTP) {
+      Integer injectsPerAttackPattern) {
     if (knownInjectorContracts == null
         || platformArchitecturePairs == null
         || platformArchitecturePairs.isEmpty()) {
@@ -269,7 +212,7 @@ public class InjectAssistantService {
                   architecture.equals(ic.getPayload().getExecutionArch().name());
               return hasPlatforms && hasArchitecture;
             })
-        .limit(injectsPerTTP)
+        .limit(injectsPerAttackPattern)
         .toList();
   }
 
@@ -280,17 +223,18 @@ public class InjectAssistantService {
    * @param attackPattern the attack pattern to match against the injector contracts
    * @param platformArchitecturePairs the list of platform-architecture pairs to filter the
    *     contracts
-   * @param injectsPerTTP the maximum number of injector contracts to return
+   * @param injectsPerAttackPattern the maximum number of injector contracts to return
    * @return a list of InjectorContract objects that match the search criteria
    */
   private List<InjectorContract> findOrSearchInjectorContract(
       List<InjectorContract> knownInjectorContracts,
       AttackPattern attackPattern,
       List<String> platformArchitecturePairs,
-      Integer injectsPerTTP) {
+      Integer injectsPerAttackPattern) {
     // Find in existing list of InjectorContracts
     List<InjectorContract> existingInjectorContract =
-        findInjectorContracts(knownInjectorContracts, platformArchitecturePairs, injectsPerTTP);
+        findInjectorContracts(
+            knownInjectorContracts, platformArchitecturePairs, injectsPerAttackPattern);
     if (!existingInjectorContract.isEmpty()) {
       return existingInjectorContract;
     }
@@ -298,28 +242,28 @@ public class InjectAssistantService {
     // Else find from DB
     return this.injectorContractRepositoryHelper
         .searchInjectorContractsByAttackPatternAndEnvironment(
-            attackPattern.getExternalId(), platformArchitecturePairs, injectsPerTTP);
+            attackPattern.getExternalId(), platformArchitecturePairs, injectsPerAttackPattern);
   }
 
   private record ContractResultForAssetGroup(
       List<InjectorContract> injectorContracts, String unmatchedPlatformArchitecture) {}
 
   /**
-   * Get the injector contracts for a specific asset group and TTP
+   * Get the injector contracts for a specific asset group and AttackPattern
    *
    * @param assetsFromGroup the assets related to group for which the injector contracts need to be
    *     found
    * @param attackPattern the attack pattern for which the injector contracts need to match
-   * @param injectsPerTTP the maximum number of injector contracts to return
+   * @param injectsPerAttackPattern the maximum number of injector contracts to return
    * @param knownInjectorContracts the list of already found injector contracts to search from
    * @return a ContractResultForAssetGroup containing the injector contracts that successfully
    *     matched for the asset group. and the most common platform-architecture pairs within the
    *     asset group for which no matching injector contract was found.
    */
-  private ContractResultForAssetGroup getInjectorContractsForAssetGroupAndTTP(
+  private ContractResultForAssetGroup getInjectorContractsForAssetGroupAndAttackPattern(
       List<Endpoint> assetsFromGroup,
       AttackPattern attackPattern,
-      Integer injectsPerTTP,
+      Integer injectsPerAttackPattern,
       List<InjectorContract> knownInjectorContracts) {
     String unmatchedPlatformArchitecture = "";
 
@@ -338,7 +282,7 @@ public class InjectAssistantService {
             knownInjectorContracts,
             attackPattern,
             groupedAssets.keySet().stream().toList(),
-            injectsPerTTP);
+            injectsPerAttackPattern);
     if (!injectorContracts.isEmpty()) {
       return new ContractResultForAssetGroup(injectorContracts, unmatchedPlatformArchitecture);
     }
@@ -353,7 +297,10 @@ public class InjectAssistantService {
     // Try to find injectors contract for the most common group
     List<InjectorContract> injectorContractsForGroup =
         findOrSearchInjectorContract(
-            knownInjectorContracts, attackPattern, List.of(mostCommonPlatformArch), injectsPerTTP);
+            knownInjectorContracts,
+            attackPattern,
+            List.of(mostCommonPlatformArch),
+            injectsPerAttackPattern);
     if (injectorContractsForGroup.isEmpty()) {
       unmatchedPlatformArchitecture = mostCommonPlatformArch;
     }
@@ -365,15 +312,15 @@ public class InjectAssistantService {
    *
    * @param endpoints the list of endpoints to process
    * @param attackPattern the attack pattern for which the injects are created
-   * @param injectsPerTTP the maximum number of injects to create for each TTP
+   * @param injectsPerAttackPattern the maximum number of injects to create for each AttackPattern
    * @param contractInjectMap a map to store the injector contracts and their corresponding injects
    * @param manualInjectMap a map to store manual injects based on platform-architecture pairs
    * @param knownInjectorContracts the list of already known injector contracts
    */
-  private void handleEndpoints(
+  public void handleEndpoints(
       List<Endpoint> endpoints,
       AttackPattern attackPattern,
-      Integer injectsPerTTP,
+      Integer injectsPerAttackPattern,
       Map<InjectorContract, Inject> contractInjectMap,
       Map<String, Inject> manualInjectMap,
       List<InjectorContract> knownInjectorContracts) {
@@ -381,7 +328,8 @@ public class InjectAssistantService {
       return;
     }
     ContractResultForEndpoints endpointResults =
-        getInjectorContractForAssetsAndTTP(attackPattern, injectsPerTTP, endpoints);
+        getInjectorContractForAssetsAndAttackPattern(
+            attackPattern, injectsPerAttackPattern, endpoints);
 
     // Add matched contracts
     endpointResults.contractEndpointsMap.forEach(
@@ -416,15 +364,15 @@ public class InjectAssistantService {
    *
    * @param assetsFromGroupMap Map of assetGroups with their list of endpoints
    * @param attackPattern the attack pattern for which the injects are created
-   * @param injectsPerTTP the maximum number of injects to create for each TTP
+   * @param injectsPerAttackPattern the maximum number of injects to create for each AttackPattern
    * @param contractInjectMap a map to store the injector contracts and their corresponding inject
    * @param manualInjectMap a map to store manual injects based on platform-architecture pairs
    * @param knownInjectorContracts the list of already known injector contracts
    */
-  private void handleAssetGroups(
+  public void handleAssetGroups(
       Map<AssetGroup, List<Endpoint>> assetsFromGroupMap,
       AttackPattern attackPattern,
-      Integer injectsPerTTP,
+      Integer injectsPerAttackPattern,
       Map<InjectorContract, Inject> contractInjectMap,
       Map<String, Inject> manualInjectMap,
       List<InjectorContract> knownInjectorContracts) {
@@ -432,8 +380,8 @@ public class InjectAssistantService {
       List<Endpoint> assetsFromGroup = assetsFromGroupMap.get(group);
 
       ContractResultForAssetGroup result =
-          getInjectorContractsForAssetGroupAndTTP(
-              assetsFromGroup, attackPattern, injectsPerTTP, knownInjectorContracts);
+          getInjectorContractsForAssetGroupAndAttackPattern(
+              assetsFromGroup, attackPattern, injectsPerAttackPattern, knownInjectorContracts);
 
       result.injectorContracts.forEach(
           contract -> {
@@ -462,207 +410,6 @@ public class InjectAssistantService {
   }
 
   /**
-   * Generates injects based on the provided attack pattern ID, endpoints, asset groups, and the
-   * number of injects to create for each TTP.
-   *
-   * @param attackPatternId the internal ID of the attack pattern to generate injects for
-   * @param endpoints the list of endpoints to consider for the injects
-   * @param assetsFromGroupMap the list of asset groups to consider for the injects
-   * @param injectsPerTTP the maximum number of injects to create for each TTP
-   * @return a list of generated injects
-   */
-  private Set<Inject> generateInjectsByTTP(
-      String attackPatternId,
-      List<Endpoint> endpoints,
-      Map<AssetGroup, List<Endpoint>> assetsFromGroupMap,
-      Integer injectsPerTTP)
-      throws UnprocessableContentException {
-
-    Set<Inject> injects;
-
-    // Check if attack pattern exist
-    AttackPattern attackPattern = attackPatternService.findById(attackPatternId);
-
-    // Otherwise, We try computing the best case (with all possible platforms and architecture)
-    injects =
-        buildInjectsForAllPlatformAndArchCombinations(
-            endpoints, new ArrayList<>(assetsFromGroupMap.keySet()), injectsPerTTP, attackPattern);
-
-    if (!injects.isEmpty()) {
-      return injects;
-    }
-
-    // Otherwise, process for all endpoints and all assetgroups to find injector contract that match
-    // with TTP and platforms/architectures
-    Map<InjectorContract, Inject> contractInjectMap = new HashMap<>();
-    Map<String, Inject> manualInjectMap = new HashMap<>();
-    List<InjectorContract> knownInjectorContracts = new ArrayList<>();
-
-    handleEndpoints(
-        endpoints,
-        attackPattern,
-        injectsPerTTP,
-        contractInjectMap,
-        manualInjectMap,
-        knownInjectorContracts);
-    handleAssetGroups(
-        assetsFromGroupMap,
-        attackPattern,
-        injectsPerTTP,
-        contractInjectMap,
-        manualInjectMap,
-        knownInjectorContracts);
-
-    injects =
-        Stream.concat(contractInjectMap.values().stream(), manualInjectMap.values().stream())
-            .collect(Collectors.toSet());
-    if (injects.isEmpty()) {
-      throw new UnprocessableContentException("No target found");
-    }
-    return injects;
-  }
-
-  // -- SCENARIO FROM STIX IMPORT : GENERATION INJECTS --
-
-  /**
-   * Generates injects for the given scenario and set of attack patterns, without considering asset
-   * groups or endpoint platform/architecture.
-   *
-   * <p>This method assumes no platform or architecture constraints and tries to generate injects
-   * using any compatible injector contract, or falls back to a generic manual inject.
-   *
-   * @param scenario the scenario to which the injects belong
-   * @param attackPatterns the set of attack patterns (AttackPatterns) to generate injects for
-   * @param injectsPerAttackPattern the number of injects to generate per AttackPattern
-   * @return the list of created and saved injects
-   */
-  public Set<Inject> generateInjectsByAttackPatternsWithoutAssetGroups(
-      Scenario scenario, Set<AttackPattern> attackPatterns, Integer injectsPerAttackPattern) {
-    Set<Inject> injects =
-        attackPatterns.stream()
-            .flatMap(
-                attackPattern ->
-                    buildInjectsForAnyPlatformAndArchitecture(
-                        injectsPerAttackPattern, attackPattern)
-                        .stream())
-            .peek(inject -> inject.setScenario(scenario))
-            .collect(Collectors.toSet());
-    return this.injectRepository.saveAll(injects).stream().collect(Collectors.toSet());
-  }
-
-  /**
-   * Generates injects for the given scenario and attack patterns, using the specified asset groups
-   * and their endpoints to guide platform and architecture selection.
-   *
-   * @param scenario the scenario to which the injects belong
-   * @param attackPatterns the set of attack patterns (AttackPatterns) to generate injects for
-   * @param injectsPerAttackPattern the number of injects to generate per AttackPattern
-   * @param assetsFromGroupMap a mapping of asset groups to their associated endpoints
-   * @return the list of created and saved injects
-   * @throws UnsupportedOperationException if inject creation fails due to unprocessable content
-   */
-  public Set<Inject> generateInjectsByAttackPatternsWithAssetGroups(
-      Scenario scenario,
-      Set<AttackPattern> attackPatterns,
-      Integer injectsPerAttackPattern,
-      Map<AssetGroup, List<Endpoint>> assetsFromGroupMap) {
-    Set<Inject> injects = new HashSet<>();
-
-    for (AttackPattern attackPattern : attackPatterns) {
-      try {
-        Set<Inject> injectsToAdd =
-            this.generateInjectsForSingleAttackPatternWithAssetGroups(
-                attackPattern, assetsFromGroupMap, injectsPerAttackPattern);
-        injectsToAdd.forEach(inject -> inject.setScenario(scenario));
-        injects.addAll(injectsToAdd);
-      } catch (UnprocessableContentException e) {
-        throw new UnsupportedOperationException(e);
-      }
-    }
-    return this.injectRepository.saveAll(injects).stream().collect(Collectors.toSet());
-  }
-
-  /**
-   * Generates injects for a single attack pattern (AttackPattern), based on the provided asset
-   * groups and their associated endpoints.
-   *
-   * <p>First attempts to use injector contracts that support all required platform-architecture
-   * pairs. If not found, performs a deeper search across endpoints and asset groups for matching
-   * contracts.
-   *
-   * @param attackPattern the attack pattern to generate injects for
-   * @param assetsFromGroupMap a mapping of asset groups to their associated endpoints
-   * @param injectsPerAttackPattern the number of injects to generate
-   * @return the list of injects generated for the given attack pattern
-   * @throws UnprocessableContentException if no valid inject configuration can be found
-   */
-  private Set<Inject> generateInjectsForSingleAttackPatternWithAssetGroups(
-      AttackPattern attackPattern,
-      Map<AssetGroup, List<Endpoint>> assetsFromGroupMap,
-      Integer injectsPerAttackPattern)
-      throws UnprocessableContentException {
-
-    // Computing best case (with all possible platforms and architecture)
-    List<Endpoint> NO_ENDPOINTS = new ArrayList<>();
-    Set<Inject> bestCaseInjects =
-        buildInjectsForAllPlatformAndArchCombinations(
-            NO_ENDPOINTS,
-            new ArrayList<>(assetsFromGroupMap.keySet()),
-            injectsPerAttackPattern,
-            attackPattern);
-
-    if (!bestCaseInjects.isEmpty()) {
-      return bestCaseInjects;
-    }
-
-    // Otherwise, process for all endpoints and all assetgroups to find injector contract that
-    // match with AttackPattern and platforms/architectures
-    Map<InjectorContract, Inject> contractInjectMap = new HashMap<>();
-    Map<String, Inject> manualInjectMap = new HashMap<>();
-    List<InjectorContract> knownInjectorContracts = new ArrayList<>();
-
-    handleAssetGroups(
-        assetsFromGroupMap,
-        attackPattern,
-        injectsPerAttackPattern,
-        contractInjectMap,
-        manualInjectMap,
-        knownInjectorContracts);
-
-    return Stream.concat(contractInjectMap.values().stream(), manualInjectMap.values().stream())
-        .collect(Collectors.toSet());
-  }
-
-  /**
-   * Attempts to generate injects for a given attack pattern (AttackPattern) without restricting to
-   * specific platforms or architectures.
-   *
-   * <p>If any compatible injector contracts exist, they are used. Otherwise, a generic manual
-   * inject is created using default values "ANY" for platform and architecture.
-   *
-   * @param injectsPerAttackPattern the number of injects to generate
-   * @param attackPattern the attack pattern to generate injects for
-   * @return the list of generated injects
-   */
-  private List<Inject> buildInjectsForAnyPlatformAndArchitecture(
-      Integer injectsPerAttackPattern, AttackPattern attackPattern) {
-    List<InjectorContract> injectorContracts =
-        this.injectorContractRepositoryHelper.searchInjectorContractsByAttackPatternAndEnvironment(
-            attackPattern.getExternalId(), emptyList(), injectsPerAttackPattern);
-
-    if (!injectorContracts.isEmpty()) {
-      return injectorContracts.stream()
-          .map(
-              ic ->
-                  buildTechnicalInjectFromInjectorContract(
-                      ic, attackPattern.getExternalId(), attackPattern.getName()))
-          .toList();
-    }
-    return List.of(
-        buildManualInject(attackPattern.getExternalId(), "[any platform]", "[any architecture]"));
-  }
-
-  /**
    * Attempts to generate injects using injector contracts that support all required
    * platform-architecture combinations for the given attack pattern.
    *
@@ -675,7 +422,7 @@ public class InjectAssistantService {
    * @param attackPattern the attack pattern to generate injects for
    * @return the list of injects, or an empty list if no contracts matched
    */
-  private Set<Inject> buildInjectsForAllPlatformAndArchCombinations(
+  public Set<Inject> buildInjectsForAllPlatformAndArchCombinations(
       List<Endpoint> endpoints,
       List<AssetGroup> assetGroups,
       Integer injectsPerAttackPattern,
@@ -696,45 +443,5 @@ public class InjectAssistantService {
               return inject;
             })
         .collect(Collectors.toSet());
-  }
-
-  // -- Vulnerabilities --
-
-  /**
-   * Generates injects for the given scenario and vulnerabilities
-   *
-   * @param scenario the scenario to which the injects belong
-   * @param vulnerabilityIds the set of Vulnerabilities (Cves) to generate injects for
-   * @param injectsPerVulnerability the number of injects to generate per Vulnerability
-   * @return the list of created and saved injects
-   */
-  public Set<Inject> generateInjectsByVulnerabilities(
-      Scenario scenario, Set<Cve> vulnerabilityIds, int injectsPerVulnerability) {
-    Set<Inject> injects =
-        vulnerabilityIds.stream()
-            .flatMap(
-                vulnerability ->
-                    buildInjectsByVulnerability(vulnerability, injectsPerVulnerability).stream())
-            .peek(inject -> inject.setScenario(scenario))
-            .collect(Collectors.toSet());
-    return this.injectRepository.saveAll(injects).stream().collect(Collectors.toSet());
-  }
-
-  private Set<Inject> buildInjectsByVulnerability(
-      Cve vulnerability, Integer injectsPerVulnerability) {
-    Set<InjectorContract> injectorContracts =
-        this.injectorContractRepository.findInjectorContractsByVulnerabilityId(
-            vulnerability.getExternalId(), injectsPerVulnerability);
-
-    if (!injectorContracts.isEmpty()) {
-      return injectorContracts.stream()
-          .map(
-              ic ->
-                  buildTechnicalInjectFromInjectorContract(
-                      ic, vulnerability.getExternalId(), vulnerability.getCisaVulnerabilityName()))
-          .collect(Collectors.toSet());
-    }
-    return Set.of(
-        buildManualInject(vulnerability.getExternalId(), "[any platform]", "[any architecture]"));
   }
 }
